fix: use setFirstName/setLastName instead of setName for Contributor
Fixed bug in user creation where the code attempted to call the non-existent
setName() method on Contributor entity. The Contributor entity uses firstName
and lastName fields separately, not a single name field.

Changed:
- setName(trim($firstName.' '.$lastName))
To:
- setFirstName($firstName)->setLastName($lastName)

This fixes the error: 'Attempted to call an undefined method named "setName"
of class "App\Entity\Contributor"' when creating a new user.